moving some code
